Address help review follow-ups
Make bare invocations consistently emit preflight guidance even when leading
configuration flags are present, and update the affected tests.

Move daemon execution helpers into `runner_glue` so `lib.rs` stays below the
repository file-size limit. Also align `insta` with the workspace dependency,
rerun manpage generation when `src/help.rs` changes, and refresh the roadmap
and execplan notes for the shipped locale/help architecture.
